+++ lib/widgets/get_images.dart
@@ -27,25 +27,25 @@ class GetImage extends StatelessWidget {
   final Color loadingBgColor;
   final Widget? errorWidget;
 
-  const GetImage({super.key,
-    this.imagePath,
-    this.width = Siz.profileImageSize,
-    this.height = Siz.profileImageSize,
-    this.fit = BoxFit.cover,
-    this.radius,
-    this.isAssets = false,
-    this.imageColor,
-    this.borderRadius,
-    this.loadingColor,
-    this.onTap,
-    this.backgroundDecoration,
-    this.pageController,
-    this.onPageChanged,
-    this.appBar,
-    this.imageLoadingDelay = const Duration(milliseconds: 100),
-    this.loadingBgColor = Colors.transparent,
-    this.errorWidget
-  });
+  const GetImage(
+      {super.key,
+      this.imagePath,
+      this.width = Siz.profileImageSize,
+      this.height = Siz.profileImageSize,
+      this.fit = BoxFit.cover,
+      this.radius,
+      this.isAssets = false,
+      this.imageColor,
+      this.borderRadius,
+      this.loadingColor,
+      this.onTap,
+      this.backgroundDecoration,
+      this.pageController,
+      this.onPageChanged,
+      this.appBar,
+      this.imageLoadingDelay = const Duration(milliseconds: 100),
+      this.loadingBgColor = Colors.transparent,
+      this.errorWidget});
 
   static String defaultImage = 'assets/default.png';
   static bool defImageIsAsset = true;
@@ -56,96 +56,106 @@ class GetImage extends StatelessWidget {
     if (imagePath == null) {
       isAsset = true;
     }
-    return GestureDetector(
-      onTap: onTap ??
-          (Static.defaultImageClick
-              ? () {
-                  pSetRout(
-                    page: () => FullPhotoView(
-                      images: [imagePath ?? defaultImage],
-                      isAsset: isAsset,
-                      isSingleImage: true,
-                      backgroundDecoration: backgroundDecoration,
-                      onPageChanged: onPageChanged,
-                      pageController: pageController,
-                      appBar: appBar,
-                    ),
-                  );
-                }
-              : null),
-      child: ClipRRect(
-        borderRadius:
-            borderRadius ?? BorderRadius.circular(radius ?? Siz.defaultRadius),
-        child: isAsset
-            ? SizedBox(
-                height: height,
-                width: width,
-                child: Image.asset(
-                  imagePath ?? defaultImage,
-                  color: imageColor,
-                  width: width,
+    return SizedBox(
+      height: height,
+      width: width,
+      child: GestureDetector(
+        onTap: onTap ??
+            (Static.defaultImageClick
+                ? () {
+                    pSetRout(
+                      page: () => FullPhotoView(
+                        images: [imagePath ?? defaultImage],
+                        isAsset: isAsset,
+                        isSingleImage: true,
+                        backgroundDecoration: backgroundDecoration,
+                        onPageChanged: onPageChanged,
+                        pageController: pageController,
+                        appBar: appBar,
+                      ),
+                    );
+                  }
+                : null),
+        child: ClipRRect(
+          borderRadius: borderRadius ?? BorderRadius.circular(radius ?? Siz.defaultRadius),
+          child:
+          isAsset
+              ?
+          SizedBox(
                   height: height,
-                  fit: fit,
-                ),
-              )
-            : buildFutureBuilderImage(),
+                  width: width,
+                  child: Image.asset(
+                    imagePath ?? defaultImage,
+                    color: imageColor,
+                    width: width,
+                    height: height,
+                    fit: fit,
+                  ),
+                )
+              :
+          buildFutureBuilderImage(),
+        ),
       ),
     );
   }
 
-  FutureBuilder<dynamic> buildFutureBuilder() {
-    return FutureBuilder(
-            future: Future.delayed(imageLoadingDelay),
-            builder: (context, asyncSnapshot) {
-              if(asyncSnapshot.connectionState != ConnectionState.done){
-                return Container(
-                  height: height,
-                  width: width,
-                  decoration: pBoxDecoration(
-                    borderRadius: borderRadius ??
-                        BorderRadius.circular(radius ?? Siz.defaultRadius),
-                  ),
-                  child: LoadingPro(
-                    valueColor: loadingColor ?? Clr.colorPrimary,
-                  ),
-                );
-              }
-              return CachedNetworkImage(
-                  imageUrl: imagePath ?? defaultImage,
-                  imageBuilder: (context, imageProvider) => Container(
-                    height: height,
-                    width: width,
-                    decoration: BoxDecoration(
-                      image: DecorationImage(
-                        image: imageProvider,
-                        fit: fit,
-                      ),
-                    ),
-                  ),
-                  placeholder: (context, url) => Container(
-                    height: height,
-                    width: width,
-                    decoration: pBoxDecoration(
-                      borderRadius: borderRadius ??
-                          BorderRadius.circular(radius ?? Siz.defaultRadius),
-                    ),
-                    child: LoadingPro(
-                      valueColor: loadingColor ?? Clr.colorPrimary,
-                    ),
-                  ),
-                  errorWidget: (context, url, error) => Container(
-                    height: height,
-                    width: width,
-                    decoration: pBoxDecoration(
-                      borderRadius: borderRadius ??
-                          BorderRadius.circular(radius ?? Siz.defaultRadius),
-                    ),
-                    child: Image.asset(defaultImage, width: width, height: height,fit: fit,),
-                  ),
-                );
-            }
-          );
-  }
+  // FutureBuilder<dynamic> buildFutureBuilder() {
+  //   return FutureBuilder(
+  //       future: Future.delayed(imageLoadingDelay),
+  //       builder: (context, asyncSnapshot) {
+  //         if (asyncSnapshot.connectionState != ConnectionState.done) {
+  //           return Container(
+  //             height: height,
+  //             width: width,
+  //             decoration: pBoxDecoration(
+  //               borderRadius: borderRadius ??
+  //                   BorderRadius.circular(radius ?? Siz.defaultRadius),
+  //             ),
+  //             child: LoadingPro(
+  //               valueColor: loadingColor ?? Clr.colorPrimary,
+  //             ),
+  //           );
+  //         }
+  //         return CachedNetworkImage(
+  //           imageUrl: imagePath ?? defaultImage,
+  //           imageBuilder: (context, imageProvider) => Container(
+  //             height: height,
+  //             width: width,
+  //             decoration: BoxDecoration(
+  //               image: DecorationImage(
+  //                 image: imageProvider,
+  //                 fit: fit,
+  //               ),
+  //             ),
+  //           ),
+  //           placeholder: (context, url) => Container(
+  //             height: height,
+  //             width: width,
+  //             decoration: pBoxDecoration(
+  //               borderRadius: borderRadius ??
+  //                   BorderRadius.circular(radius ?? Siz.defaultRadius),
+  //             ),
+  //             child: LoadingPro(
+  //               valueColor: loadingColor ?? Clr.colorPrimary,
+  //             ),
+  //           ),
+  //           errorWidget: (context, url, error) => Container(
+  //             height: height,
+  //             width: width,
+  //             decoration: pBoxDecoration(
+  //               borderRadius: borderRadius ??
+  //                   BorderRadius.circular(radius ?? Siz.defaultRadius),
+  //             ),
+  //             child: Image.asset(
+  //               defaultImage,
+  //               width: width,
+  //               height: height,
+  //               fit: fit,
+  //             ),
+  //           ),
+  //         );
+  //       });
+  // }
 
   Widget buildFutureBuilderImage() {
     return FutureBuilder(
@@ -166,7 +176,24 @@ class GetImage extends StatelessWidget {
               );
             },
             progressIndicatorBuilder: (context, url, downloadProgress) {
-              double progress = downloadProgress.progress != null && downloadProgress.progress! > 0.0 && downloadProgress.progress! < 1.0? downloadProgress.progress! : 0.0;
+              double progress = downloadProgress.progress != null && downloadProgress.progress! > 0.0 && downloadProgress.progress! < 1.0 ? downloadProgress.progress! : 0.0;
+              if (progress < 0.1) {
+                return Center(
+                  child: Container(
+                    height: height,
+                      width: width,
+                    decoration: BoxDecoration(
+                      shape: BoxShape.circle,
+                      // borderRadius: (borderRadius ?? BorderRadius.circular(radius ?? Siz.defaultRadius)),
+                    ),
+                    child: Center(
+                      child: LoadingPro(
+                        valueColor: loadingColor ?? Clr.colorPrimary,
+                      ),
+                    ),
+                  ),
+                );
+              }
               return Center(
                 child: CircleAvatar(
                   backgroundColor: loadingBgColor,
@@ -182,23 +209,29 @@ class GetImage extends StatelessWidget {
                           color: loadingColor,
                         ),
                       ),
-                      Txt('${(progress * 100).toStringAsFixed(0)}%', fontSize: 12, textColor: Clr.colorPrimary),
+                      Txt('${(progress * 100).toStringAsFixed(0)}%',
+                          fontSize: 12, textColor: Clr.colorPrimary),
                     ],
                   ),
                 ),
               );
             },
-            errorWidget: (context, url, error) => errorWidget??Container(
-              height: height,
-              width: width,
-              decoration: pBoxDecoration(
-                borderRadius: borderRadius ??
-                    BorderRadius.circular(radius ?? Siz.defaultRadius),
-              ),
-              child: Image.asset(defaultImage, width: width, height: height,fit: fit,),
-            ),
+            errorWidget: (context, url, error) =>
+                errorWidget ??
+                Container(
+                  height: height,
+                  width: width,
+                  decoration: pBoxDecoration(
+                    borderRadius: borderRadius ?? BorderRadius.circular(radius ?? Siz.defaultRadius),
+                  ),
+                  child: Image.asset(
+                    defaultImage,
+                    width: width,
+                    height: height,
+                    fit: fit,
+                  ),
+                ),
           );
-        }
-    );
+        });
   }
 }